Fix up rebase from dashboard refactor
