+++ apps/dashboard/src/app/article/write-form.tsx
@@ -1,5 +1,8 @@
-import { createCheckboxInput, createFileInput, createRichTextInput, createTagInput } from "@/app/form"
+import { createCheckboxInput } from "@/components/forms/CheckboxInput"
+import { createFileInput } from "@/components/forms/FileInput"
 import { useFormBuilder } from "@/components/forms/Form"
+import { createRichTextInput } from "@/components/forms/RichTextInput"
+import { createTagInput } from "@/components/forms/TagInput"
 import { createTextInput } from "@/components/forms/TextInput"
 import { ArticleSchema, type ArticleWrite, ArticleWriteSchema } from "@dotkomonline/types"
 import { useTagsAllQuery } from "src/modules/article/queries/use-tags-all-query"
@@ -68,12 +71,10 @@ export const useArticleWriteForm = ({
       }),
       excerpt: createRichTextInput({
         label: "Ingress",
-        markdown: "",
         required: true,
       }),
       content: createRichTextInput({
         label: "Innhold",
-        markdown: "",
         required: true,
       }),
     },
